feat: add authorization to automation API
see https://gravitee.atlassian.net/browse/GKO-1428

+++ gravitee-apim-rest-api/gravitee-apim-rest-api-automation/gravitee-apim-rest-api-automation-rest/src/main/java/io/gravitee/apim/rest/api/automation/resource/ApiSubscriptionResource.java
@@ -15,8 +15,6 @@
  */
 package io.gravitee.apim.rest.api.automation.resource;
 
-import io.gravitee.apim.core.api.crud_service.ApiCrudService;
-import io.gravitee.apim.core.api.model.Api;
 import io.gravitee.apim.core.application.crud_service.ApplicationCrudService;
 import io.gravitee.apim.core.plan.crud_service.PlanCrudService;
 import io.gravitee.apim.core.plan.model.Plan;
@@ -48,13 +46,10 @@ import lombok.extern.slf4j.Slf4j;
  * @author GraviteeSource Team
  */
 @Slf4j
-public class SubscriptionResource extends AbstractResource {
+public class ApiSubscriptionResource extends AbstractResource {
 
     @Inject
     private SubscriptionCrudService subscriptionCrudService;
-
-    @Inject
-    private ApiCrudService apiCrudService;
 
     @Inject
     private ApplicationCrudService applicationCrudService;
@@ -62,28 +57,30 @@ public class SubscriptionResource extends AbstractResource {
     @Inject
     private PlanCrudService planCrudService;
 
-    @PathParam("hrid")
-    private String hrid;
-
     @GET
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
     @Permissions({ @Permission(value = RolePermission.API_SUBSCRIPTION, acls = { RolePermissionAction.READ }) })
-    public Response getSubscriptionByHRID(@QueryParam("legacy") boolean legacy) {
+    public Response getSubscriptionByHRID(
+        @PathParam("apiHrid") String apiHrid,
+        @PathParam("hrid") String hrid,
+        @QueryParam("legacy") boolean legacy
+    ) {
         var executionContext = GraviteeContext.getExecutionContext();
         try {
+            String subscriptionId = legacy ? hrid : IdBuilder.builder(executionContext, apiHrid).withExtraId(hrid).buildId();
+            SubscriptionEntity subscriptionEntity = subscriptionCrudService.get(subscriptionId);
+
+            if (legacy && !subscriptionEntity.getApiId().equals(apiHrid)) {
+                throw new SubscriptionNotFoundException(apiHrid);
+            }
+
             SubscriptionState subscriptionState = new SubscriptionState();
+            subscriptionState.setId(subscriptionId);
             subscriptionState.setHrid(hrid);
             subscriptionState.setEnvironmentId(executionContext.getEnvironmentId());
             subscriptionState.setOrganizationId(executionContext.getOrganizationId());
-
-            SubscriptionEntity subscriptionEntity = subscriptionCrudService.get(
-                legacy ? hrid : IdBuilder.builder(executionContext, hrid).buildId()
-            );
-            subscriptionState.setId(subscriptionEntity.getId());
-
-            Api api = apiCrudService.get(subscriptionEntity.getApiId());
-            subscriptionState.setApiHrid(api.getHrid());
+            subscriptionState.setApiHrid(apiHrid);
 
             BaseApplicationEntity application = applicationCrudService.findById(
                 subscriptionEntity.getApplicationId(),
@@ -100,22 +97,30 @@ public class SubscriptionResource extends AbstractResource {
             );
             return Response.ok(subscriptionState).build();
         } catch (SubscriptionNotFoundException e) {
+            log.debug("Subscription not found for hrid: {}, apiHrid {}, operation: get", hrid, apiHrid);
             throw new HRIDNotFoundException(hrid);
         }
     }
 
     @DELETE
     @Permissions({ @Permission(value = RolePermission.API_SUBSCRIPTION, acls = RolePermissionAction.DELETE) })
-    public Response deleteSubscriptionByHrid(@QueryParam("legacy") boolean legacy) {
+    public Response deleteSubscriptionByHrid(
+        @PathParam("apiHrid") String apiHrid,
+        @PathParam("hrid") String hrid,
+        @QueryParam("legacy") boolean legacy
+    ) {
         var executionContext = GraviteeContext.getExecutionContext();
-
         try {
-            SubscriptionEntity subscriptionEntity = subscriptionCrudService.get(
-                legacy ? hrid : IdBuilder.builder(executionContext, hrid).buildId()
-            );
+            String subscriptionId = legacy ? hrid : IdBuilder.builder(executionContext, apiHrid).withExtraId(hrid).buildId();
+            SubscriptionEntity subscriptionEntity = subscriptionCrudService.get(subscriptionId);
+
+            if (legacy && !subscriptionEntity.getApiId().equals(apiHrid)) {
+                throw new SubscriptionNotFoundException(apiHrid);
+            }
+
             subscriptionCrudService.delete(subscriptionEntity.getId());
         } catch (SubscriptionNotFoundException e) {
-            log.warn("SharedPolicyGroup not found for hrid: {}, operation: delete", hrid);
+            log.debug("Subscription not found for hrid: {}, apiHrid {}, operation: delete", hrid, apiHrid);
             throw new HRIDNotFoundException(hrid);
         }
         return Response.noContent().build();
